requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
typing
joblib
scikit-learn
python-dotenv